Update draft-ietf-scim-cursor-pagination-04.md
Updates for Other comments:

In the Abstract and Introduction, a nit:
“is already well established” should not be hyphenated, as it’s not a modifier
(whereas “a well-established pagination pattern” is correctly hyphenated).
Resolution: removed hyphen

— Section 2 —

   The following table describes the URL pagination parameters requests
   for using cursor-based pagination:

I think the word “requests” is extra and should be removed (or perhaps “using”
should be replaced by “requesting”).

Resolution: replaced "requests for using" with "for requesting"

In the second table you say “Use of previousCursor is OPTIONAL.”  That seems to
say that using it in a subsequent request is optional, which is already said in
the previous sentence.  I think this one should say, “Returning previousCursor
is OPTIONAL.”

Resolution: replaced “Use of previousCursor is OPTIONAL.” with "Returning previousCursor
is OPTIONAL."